moved DolphinSessionListener to public API
