+++ platform/dolphin-platform-server/src/test/java/com/canoo/dolphin/server/config/ConfigurationFileLoaderTest.java
@@ -15,7 +15,6 @@
  */
 package com.canoo.dolphin.server.config;
 
-import com.canoo.dolphin.server.impl.ConfigurationFileLoader;
 import org.testng.annotations.Test;
 
 import java.util.logging.Level;
